Added Search/Sort functionality
Added search functionality to WAR Accomplishments page.  This allows for users to search for projects via entering in the name of the project into the text field.  

Added sort functionality to WAR Accomplishments page. This allows for user to sort their results based on Project Name or Week Ending Date.